+++ src/ska_mid_dish_manager/devices/DishManagerDS.py
@@ -136,6 +136,10 @@ class DishManager(SKAController):
             ("TrackStop", "track_stop_cmd"),
             ("ConfigureBand1", "configure_band_cmd"),
             ("ConfigureBand2", "configure_band_cmd"),
+            ("ConfigureBand3", "configure_band_cmd"),
+            ("ConfigureBand4", "configure_band_cmd"),
+            ("ConfigureBand5a", "configure_band_cmd"),
+            ("ConfigureBand5b", "configure_band_cmd"),
             ("Slew", "slew"),
             ("Scan", "scan"),
             ("TrackLoadStaticOff", "track_load_static_off"),
@@ -1580,7 +1584,7 @@ class DishManager(SKAController):
         """
         handler = self.get_command_object("ConfigureBand1")
 
-        result_code, unique_id = handler("1", synchronise)
+        result_code, unique_id = handler(Band.B1, synchronise)
         return ([result_code], [unique_id])
 
     @record_command(False)
@@ -1607,7 +1611,7 @@ class DishManager(SKAController):
         """
         handler = self.get_command_object("ConfigureBand2")
 
-        result_code, unique_id = handler("2", synchronise)
+        result_code, unique_id = handler(Band.B2, synchronise)
         return ([result_code], [unique_id])
 
     @record_command(False)
@@ -1617,7 +1621,7 @@ class DishManager(SKAController):
         doc_in="If the synchronise argument is True, the SPFRx FPGA is instructed to synchronise "
         "its internal flywheel 1PPS to the SAT-1PPS for the ADC that is applicable to the band "
         "being configured, and the band counters are reset. (Should be default to False).",
-        dtype_out=None,
+        dtype_out="DevVarLongStringArray",
         display_level=DispLevel.OPERATOR,
     )
     def ConfigureBand3(self, synchronise):  # pylint: disable=unused-argument
@@ -1627,7 +1631,10 @@ class DishManager(SKAController):
         configuration, Dish will automatically revert to the previous Dish
         mode (OPERATE or STANDBY‐FP).
         """
-        raise NotImplementedError
+        handler = self.get_command_object("ConfigureBand3")
+
+        result_code, unique_id = handler(Band.B3, synchronise)
+        return ([result_code], [unique_id])
 
     @record_command(False)
     @BaseInfoIt(show_args=True, show_kwargs=True, show_ret=True)
@@ -1636,7 +1643,7 @@ class DishManager(SKAController):
         doc_in="If the synchronise argument is True, the SPFRx FPGA is instructed to synchronise "
         "its internal flywheel 1PPS to the SAT-1PPS for the ADC that is applicable to the band "
         "being configured, and the band counters are reset. (Should be default to False).",
-        dtype_out=None,
+        dtype_out="DevVarLongStringArray",
         display_level=DispLevel.OPERATOR,
     )
     def ConfigureBand4(self, synchronise):  # pylint: disable=unused-argument
@@ -1646,7 +1653,10 @@ class DishManager(SKAController):
         configuration, Dish will automatically revert to the previous Dish
         mode (OPERATE or STANDBY‐FP).
         """
-        raise NotImplementedError
+        handler = self.get_command_object("ConfigureBand4")
+
+        result_code, unique_id = handler(Band.B4, synchronise)
+        return ([result_code], [unique_id])
 
     @record_command(False)
     @BaseInfoIt(show_args=True, show_kwargs=True, show_ret=True)
@@ -1655,7 +1665,7 @@ class DishManager(SKAController):
         doc_in="If the synchronise argument is True, the SPFRx FPGA is instructed to synchronise "
         "its internal flywheel 1PPS to the SAT-1PPS for the ADC that is applicable to the band "
         "being configured, and the band counters are reset. (Should be default to False).",
-        dtype_out=None,
+        dtype_out="DevVarLongStringArray",
         display_level=DispLevel.OPERATOR,
     )
     def ConfigureBand5a(self, synchronise):  # pylint: disable=unused-argument
@@ -1665,7 +1675,10 @@ class DishManager(SKAController):
         configuration, Dish will automatically revert to the previous Dish
         mode (OPERATE or STANDBY‐FP).
         """
-        raise NotImplementedError
+        handler = self.get_command_object("ConfigureBand5a")
+
+        result_code, unique_id = handler(Band.B5a, synchronise)
+        return ([result_code], [unique_id])
 
     @record_command(False)
     @BaseInfoIt(show_args=True, show_kwargs=True, show_ret=True)
@@ -1674,7 +1687,7 @@ class DishManager(SKAController):
         doc_in="If the synchronise argument is True, the SPFRx FPGA is instructed to synchronise "
         "its internal flywheel 1PPS to the SAT-1PPS for the ADC that is applicable to the band "
         "being configured, and the band counters are reset. (Should be default to False).",
-        dtype_out=None,
+        dtype_out="DevVarLongStringArray",
         display_level=DispLevel.OPERATOR,
     )
     def ConfigureBand5b(self, synchronise):  # pylint: disable=unused-argument
@@ -1684,7 +1697,12 @@ class DishManager(SKAController):
         configuration, Dish will automatically revert to the previous Dish
         mode (OPERATE or STANDBY‐FP).
         """
-        raise NotImplementedError
+        handler = self.get_command_object("ConfigureBand5b")
+
+        self.logger.warning("ConfigureBand5b called, but we're configuring B1 until 5B is ready.")
+        result_code, unique_id = handler(Band.B1, synchronise)
+
+        return ([result_code], [unique_id])
 
     @record_command(False)
     @command(dtype_in=None, dtype_out=None, display_level=DispLevel.OPERATOR)